Role validation in methods

+++ BookSpark/Controllers/AuthorController.cs
@@ -1,4 +1,5 @@
-﻿using BookSpark.Models.AuthorViewModels;
+﻿using BookSpark.Data.Enums;
+using BookSpark.Models.AuthorViewModels;
 using BookSpark.Models.GenreViewModels;
 using BookSpark.Services;
 using BookSpark.Services.Interfaces;
@@ -18,6 +19,10 @@ namespace BookSpark.Controllers
 
         public IActionResult Index()
         {
+            if (!User.Identity.IsAuthenticated)
+            {
+                return RedirectToAction(nameof(AuthorsError));
+            }
             var authors = authorService.GetAll().ToList();
             return View(authors);
 
@@ -25,12 +30,20 @@ namespace BookSpark.Controllers
 
         public IActionResult Add()
         {
+            if (!User.IsInRole(Roles.Admin.ToString()))
+            {
+                return RedirectToAction(nameof(AuthorsAdminError));
+            }
             return View();
         }
 
         [HttpPost]
         public IActionResult Add(AddAuthorViewModel author)
         {
+            if (!User.IsInRole(Roles.Admin.ToString()))
+            {
+                return RedirectToAction(nameof(AuthorsAdminError));
+            }
             authorService.Add(author);
 
             return RedirectToAction(nameof(Index));
@@ -38,12 +51,20 @@ namespace BookSpark.Controllers
 
         public IActionResult Delete(int id)
         {
+            if (!User.IsInRole(Roles.Admin.ToString()))
+            {
+                return RedirectToAction(nameof(AuthorsAdminError));
+            }
             authorService.Delete(id);
             return RedirectToAction(nameof(Index));
         }
 
         public IActionResult Edit(int id)
         {
+            if (!User.IsInRole(Roles.Admin.ToString()))
+            {
+                return RedirectToAction(nameof(AuthorsAdminError));
+            }
             var author = authorService.GetEditable(id);
             return View(author);
         }
@@ -51,13 +72,31 @@ namespace BookSpark.Controllers
         [HttpPost]
         public IActionResult Edit(EditAuthorViewModel author)
         {
+            if (!User.IsInRole(Roles.Admin.ToString()))
+            {
+                return RedirectToAction(nameof(AuthorsAdminError));
+            }
             authorService.Edit(author);
             return RedirectToAction(nameof(Index));
         }
         public IActionResult Detail(int id)
         {
+            if (!User.IsInRole(Roles.Admin.ToString()))
+            {
+                return RedirectToAction(nameof(AuthorsAdminError));
+            }
             var author = authorService.Get(id);
             return View(author);
+        }
+
+        public IActionResult AuthorsError()
+        {
+            return View();
+        }
+
+        public IActionResult AuthorsAdminError()
+        {
+            return View();
         }
     }
 }
